+++ test/test_base.py
@@ -17,7 +17,6 @@ import torch
 import torch.nn as nn
 
 from float8_experimental.config import CastConfig, Float8LinearConfig, ScalingType
-from float8_experimental.float8_dynamic_utils import cast_to_float8_e4m3_dynamic
 from float8_experimental.float8_linear import Float8Linear
 from float8_experimental.float8_linear_utils import (
     convert_to_float8_training,
@@ -25,11 +24,13 @@ from float8_experimental.float8_linear_utils import (
     sync_float8_amax_and_scale_history,
 )
 from float8_experimental.float8_python_api import addmm_float8_unwrapped
+from float8_experimental.float8_scaling_utils import cast_to_float8_e4m3_dynamic
 from float8_experimental.float8_tensor import (
     Float8Tensor,
     GemmInputRole,
     LinearMMConfig,
     ScaledMMConfig,
+    ToFloat8ConstrFunc,
 )
 from float8_experimental.float8_utils import (
     compute_error,
@@ -65,7 +66,7 @@ class TestFloat8Tensor(unittest.TestCase):
         for hp_dtype, lp_dtype in itertools.product(hp_dtypes, lp_dtypes):
             x1_hp = torch.randn(4, 4, dtype=hp_dtype)
             x1_s = tensor_to_scale(x1_hp, lp_dtype)
-            x2_lp = Float8Tensor.to_float8(x1_hp, x1_s, lp_dtype)
+            x2_lp = ToFloat8ConstrFunc.apply(x1_hp, x1_s, lp_dtype)
             x3_hp = x2_lp.to_original_precision()
             self.assertTrue(x3_hp.dtype == hp_dtype)
 
@@ -75,7 +76,7 @@ class TestFloat8Tensor(unittest.TestCase):
             x = torch.randn(1).requires_grad_()
             grad = torch.randn(1)
             x_s = tensor_to_scale(x, f8_dtype)
-            x_f8 = Float8Tensor.to_float8(x, x_s, f8_dtype)
+            x_f8 = ToFloat8ConstrFunc.apply(x, x_s, f8_dtype)
             x_f8_hp = x_f8.to_original_precision()
             x_f8_hp.backward(grad)
             # the gradient should be unchanged through both casts
@@ -84,7 +85,7 @@ class TestFloat8Tensor(unittest.TestCase):
     def test_split_cat(self):
         a = torch.rand(16, 16, dtype=torch.bfloat16)
         scale = tensor_to_scale(a, e4m3_dtype)
-        fp8_a = Float8Tensor.to_float8(a, scale, e4m3_dtype)
+        fp8_a = ToFloat8ConstrFunc.apply(a, scale, e4m3_dtype)
 
         splits = torch.split(fp8_a, 16)
         catted = torch.cat(splits, dim=0)
@@ -93,14 +94,14 @@ class TestFloat8Tensor(unittest.TestCase):
     def test_index_put(self):
         a = torch.rand(16, dtype=torch.bfloat16)
         scale_a = tensor_to_scale(a, torch.float8_e4m3fn)
-        fp8_a = Float8Tensor.to_float8(a, scale_a, torch.float8_e4m3fn)
+        fp8_a = ToFloat8ConstrFunc.apply(a, scale_a, torch.float8_e4m3fn)
 
         index = torch.randint(0, 15, (16,), dtype=torch.long)
 
         b = torch.rand(16, 16, dtype=torch.bfloat16)
         scale_b = tensor_to_scale(b, torch.float8_e4m3fn)
-        fp8_b = Float8Tensor.to_float8(b, scale_a, torch.float8_e4m3fn)
-        fp8_b_bad = Float8Tensor.to_float8(b, scale_b, torch.float8_e4m3fn)
+        fp8_b = ToFloat8ConstrFunc.apply(b, scale_a, torch.float8_e4m3fn)
+        fp8_b_bad = ToFloat8ConstrFunc.apply(b, scale_b, torch.float8_e4m3fn)
 
         with self.assertRaises(AssertionError):
             b[index] = fp8_a
@@ -111,7 +112,7 @@ class TestFloat8Tensor(unittest.TestCase):
     def test_copy_(self):
         a = torch.rand(16, dtype=torch.bfloat16)
         scale_a = tensor_to_scale(a, torch.float8_e4m3fn)
-        fp8_a = Float8Tensor.to_float8(a, scale_a, torch.float8_e4m3fn)
+        fp8_a = ToFloat8ConstrFunc.apply(a, scale_a, torch.float8_e4m3fn)
 
         b = torch.empty(16, dtype=torch.bfloat16)
         b.copy_(fp8_a)  # Should work
@@ -406,8 +407,8 @@ class TestScaledMM:
         a_scale = tensor_to_scale(a, input_dtype).float()
         b_scale = tensor_to_scale(b, input_dtype).float()
 
-        a_fp8 = Float8Tensor.to_float8(a, a_scale, input_dtype)
-        b_fp8 = Float8Tensor.to_float8(b, b_scale, input_dtype)
+        a_fp8 = ToFloat8ConstrFunc.apply(a, a_scale, input_dtype)
+        b_fp8 = ToFloat8ConstrFunc.apply(b, b_scale, input_dtype)
 
         out_scaled_mm = addmm_float8_unwrapped(
             a_fp8._data,
@@ -446,19 +447,21 @@ class TestScaledMM:
             ScaledMMConfig(True, False, False, False),
             ScaledMMConfig(True, False, False, False),
         )
-        a = Float8Tensor.to_float8(
+        a = ToFloat8ConstrFunc.apply(
             x_fp32,
             x_scale,
             fp8_dtype,
-            linear_mm_config=linear_config_a,
-            gemm_input_role=GemmInputRole.INPUT,
+            None,  # amax_buffer
+            linear_config_a,
+            GemmInputRole.INPUT,
         )
-        b = Float8Tensor.to_float8(
+        b = ToFloat8ConstrFunc.apply(
             x_fp32,
             x_scale,
             fp8_dtype,
-            linear_mm_config=linear_config_b,
-            gemm_input_role=GemmInputRole.WEIGHT,
+            None,  # amax_buffer
+            linear_config_b,
+            GemmInputRole.WEIGHT,
         )
         with pytest.raises(
             AssertionError,
@@ -485,11 +488,11 @@ class TestScaledMM:
         a_scale = tensor_to_scale(a, input_dtype).float()
         b_scale = tensor_to_scale(b, input_dtype).float()
 
-        a_fp8 = Float8Tensor.to_float8(
-            a, a_scale, input_dtype, gemm_input_role=GemmInputRole.INPUT
+        a_fp8 = ToFloat8ConstrFunc.apply(
+            a, a_scale, input_dtype, None, None, GemmInputRole.INPUT
         )
-        b_fp8 = Float8Tensor.to_float8(
-            b, b_scale, input_dtype, gemm_input_role=GemmInputRole.WEIGHT
+        b_fp8 = ToFloat8ConstrFunc.apply(
+            b, b_scale, input_dtype, None, None, GemmInputRole.WEIGHT
         )
 
         with pytest.raises(
@@ -505,19 +508,21 @@ class TestScaledMM:
             scaled_mm_config, scaled_mm_config, scaled_mm_config
         )
 
-        a_fp8 = Float8Tensor.to_float8(
+        a_fp8 = ToFloat8ConstrFunc.apply(
             a,
             a_scale,
             input_dtype,
-            linear_mm_config=pad_config,
-            gemm_input_role=GemmInputRole.INPUT,
+            None,  # amax_buffer
+            pad_config,
+            GemmInputRole.INPUT,
         )
-        b_fp8 = Float8Tensor.to_float8(
+        b_fp8 = ToFloat8ConstrFunc.apply(
             b,
             b_scale,
             input_dtype,
-            linear_mm_config=pad_config,
-            gemm_input_role=GemmInputRole.WEIGHT,
+            None,  # amax_buffer
+            pad_config,
+            GemmInputRole.WEIGHT,
         )
         out_padded = a_fp8 @ b_fp8
         out_padded.to(compare_type)
@@ -528,19 +533,21 @@ class TestScaledMM:
             emulated_scaled_mm_config,
             emulated_scaled_mm_config,
         )
-        a_fp8 = Float8Tensor.to_float8(
+        a_fp8 = ToFloat8ConstrFunc.apply(
             a,
             a_scale,
             input_dtype,
-            linear_mm_config=emulated_config,
-            gemm_input_role=GemmInputRole.INPUT,
+            None,  # amax_buffer
+            emulated_config,
+            GemmInputRole.INPUT,
         )
-        b_fp8 = Float8Tensor.to_float8(
+        b_fp8 = ToFloat8ConstrFunc.apply(
             b,
             b_scale,
             input_dtype,
-            linear_mm_config=emulated_config,
-            gemm_input_role=GemmInputRole.WEIGHT,
+            None,  # amax_buffer
+            emulated_config,
+            GemmInputRole.WEIGHT,
         )
         out_emualted = a_fp8 @ b_fp8
         out_emualted.to(compare_type)
@@ -694,19 +701,19 @@ class TestFloat8LinearUtils(unittest.TestCase):
 
             # Overflow caused by a too large scaling factor
             s_overflow = torch.tensor(1e9)
-            fp8_overflow = Float8Tensor.to_float8(x1_hp, s_overflow, lp_dtype)
+            fp8_overflow = ToFloat8ConstrFunc.apply(x1_hp, s_overflow, lp_dtype)
             (zero_cnt, max_cnt) = fp8_tensor_statistics(fp8_overflow, lp_dtype)
             self.assertEqual((zero_cnt, max_cnt), (0, tensor_len))
 
             # Underflow caused by a too small scaling factor
             s_underflow = torch.tensor(1e-9)
-            fp8_underflow = Float8Tensor.to_float8(x1_hp, s_underflow, lp_dtype)
+            fp8_underflow = ToFloat8ConstrFunc.apply(x1_hp, s_underflow, lp_dtype)
             (zero_cnt, max_cnt) = fp8_tensor_statistics(fp8_underflow, lp_dtype)
             self.assertEqual((zero_cnt, max_cnt), (tensor_len, 0))
 
             # Both overflow and underflow
             x2_hp = torch.cat((x1_hp * 1e9, x1_hp * 1.0, x1_hp * 1e-9), 0)
-            fp8_over_underflow = Float8Tensor.to_float8(
+            fp8_over_underflow = ToFloat8ConstrFunc.apply(
                 x2_hp, torch.tensor(1.0), lp_dtype
             )
             (zero_cnt, max_cnt) = fp8_tensor_statistics(fp8_over_underflow, lp_dtype)
